log stale migration

+++ content/auto-export.ts
@@ -482,21 +482,25 @@ export const AutoExport = new class _AutoExport { // eslint-disable-line @typesc
 
     // migration
     if (tables.includes('better-bibtex')) {
-      log.debug('autoexport migration?', await Zotero.DB.columnQueryAsync('SELECT name FROM betterbibtex."better-bibtex"'))
-      const data = await Zotero.DB.valueQueryAsync('SELECT data FROM betterbibtex."better-bibtex" WHERE name=?', ['better-bibtex.autoexport'])
-      if (data) {
-        const db = JSON.parse(data)
-        for (const ae of db.data) {
-          ae.error = ae.error || ''
-          ae.updated = ae.meta.updated
-          ae.recursive = ae.recursive || 0
-          log.debug('aedb: upgrade', ae)
-          await SQL.create(ae)
+      const collections = await Zotero.DB.columnQueryAsync('SELECT name FROM betterbibtex."better-bibtex"')
+      if (collections.includes('migrated.autoexport') && collections.includes('better-bibtex.autoexport')) {
+        flash('stale auto-export migration')
+      }
+      if (!collections.includes('migrated.autoexport')) {
+        const data = await Zotero.DB.valueQueryAsync('SELECT data FROM betterbibtex."better-bibtex" WHERE name=?', ['better-bibtex.autoexport'])
+        if (data) {
+          const db = JSON.parse(data)
+          for (const ae of db.data) {
+            ae.error = ae.error || ''
+            ae.updated = ae.meta.updated
+            ae.recursive = ae.recursive || 0
+            log.debug('aedb: upgrade', ae)
+            await SQL.create(ae)
+          }
+
+          await Zotero.DB.queryTx('DELETE FROM betterbibtex."better-bibtex" WHERE name=?', ['migrated.autoexport']) // HOW?!?!
+          await Zotero.DB.queryTx('UPDATE betterbibtex."better-bibtex" SET name = ? WHERE name = ?', ['migrated.autoexport', 'better-bibtex.autoexport'])
         }
-
-
-        await Zotero.DB.queryTx('DELETE FROM betterbibtex."better-bibtex" WHERE name=?', ['migrated.autoexport']) // HOW?!?!
-        await Zotero.DB.queryTx('UPDATE betterbibtex."better-bibtex" SET name = ? WHERE name = ?', ['migrated.autoexport', 'better-bibtex.autoexport'])
       }
     }
   }
